% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{buildData}
\alias{buildData}
\title{Build Data Object}
\usage{
buildData(Y, D, n, J = NULL, treat = NULL, segments = NULL, skip = NULL,
  model = "Model1", offset = 0)
}
\arguments{
\item{Y}{Activity series (length D*n)}

\item{D}{Number of days in study}

\item{n}{Observations per day}

\item{J}{Number of activity periods}

\item{treat}{Length of treatmens in days}

\item{segments}{Number of activity periods in each spline segment}

\item{offset}{Offset of initial time}
}
\value{
Data object
}
\description{
Build data object.
}
\details{
Build data object.
}
\author{
Simon Bonner
}

