% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/circspline.R
\name{circspline}
\alias{circspline}
\title{Fit Spline}
\usage{
circspline(data, niter, nburnin = NULL, nthin = NULL, plotinits = TRUE,
  verbose = TRUE)
}
\arguments{
\item{data}{Data object generated by buildData().}

\item{niter}{Number of iteration to run.}

\item{nburnin}{Number of burnin iterations.}

\item{nthin}{Thinning value used in plotting.}

\item{plotinits}{If TRUE then plot initial values.}

\item{verbose}{If TRUE then provide progess updates.}
}
\value{
List containing MCMC output and summarized results.
}
\description{
Fit spline model to circadian rhythm data.
}
\details{
Fit spline model to circadian rhythm data.
}
\author{
Simon Bonner
}

