% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{MCMC for Spline MOdel}
\usage{
mcmc(data, niter, inits = NULL, priors = NULL, proposals = NULL,
  updates = NULL, nadapt = 1000, ladapt = 500, radapt = 1.01,
  temper = 1, model, debug = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{Data list}

\item{niter}{Number of iterations}

\item{inits}{Initial value list}

\item{priors}{List of parameters of the prior distributions}

\item{proposals}{List of parameters of the proposal distributions}

\item{updates}{List of parameters to update (if NULL then update all)}

\item{nadapt}{Length of adapting phase}

\item{ladapt}{Iterations between adapting for multivariate parameters}

\item{temper}{Tempering value}
}
\value{
List of output
}
\description{
MCMC wrapper for spline Model 1. (Version 1)
}
\details{
This version of the MCMC algorithm augments the data with lambda
(the individual time mean counts).
}
\author{
Simon Bonner
}

