% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/priorParameters.R
\name{priorParameters}
\alias{priorParameters}
\title{Prior Parameters}
\usage{
priorParameters(data, verbose)
}
\arguments{
\item{data}{Data object generated by buildData}

\item{verbose}{If TRUE then update progress}
}
\value{
List of prior parameter values
}
\description{
Generate parameters for prior distributions.
}
\details{
These parameters are at the highest level of the hierarchy. They
are fixed by the user and are not updated.
}
\author{
Simon Bonner
}

