% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/propParameters.R
\name{propParameters}
\alias{propParameters}
\title{Proposal Parameters}
\usage{
propParameters(data, priors, verbose = TRUE)
}
\arguments{
\item{data}{Data object generated by buildData}

\item{priors}{Prior parameters list generated with priorParameters}

\item{verbose}{If TRUE then update progress}
}
\value{
List of proposal parameters
}
\description{
Define parameters of proposal distributions.
}
\details{
Define parameters of proposal distributions.
}
\author{
Simon Bonner
}

